Fixed Styling to Match Register and Login Page-Merge Ready
I believe it is fully ready to merge for real this time. Double Checked. It's a good thing we do these pull requests. Someone should look at it again still but I believe it is as good as it can get in terms of what we can do at the moment.