Store login info globally with AsyncStorage
Now, after a login or register, instead of merely passing the email and token to the Overview scene as a prop, the email and token are stored into globally accessable app storage via React Native's AsyncStorage API. They are then retrieved from AsyncStorage by the Overview scene. This means the Overview scene will be able to get that data even when navigated to via the drawer or other non-login methods.